Move autoScalers property to DefaultWorkerBehaviourConfig and remove CategorizedWorkerBehaviourConfig class (#9350)
