+++ indexing-service/src/test/java/org/apache/druid/indexing/overlord/autoscaling/SimpleProvisioningStrategyTest.java
@@ -87,7 +87,8 @@ public class SimpleProvisioningStrategyTest
     workerConfig = new AtomicReference<>(
         new DefaultWorkerBehaviorConfig(
             null,
-            autoScaler
+            autoScaler,
+            null
         )
     );
